+++ source/scene/singlethreaded/iscene_manager.h
@@ -19,7 +19,7 @@ namespace SceneModule {
     public:
 		virtual IScene* GetScene() = 0;
 		virtual void AddProcessor(IProcessor* processor) = 0;
-		virtual void Update(int dt) = 0;
+		virtual void Update(float dt) = 0;
 
 		//	processor notifications
 		//virtual void SceneChangeStarted(ISceneProcessor* value) = 0;
